requirements.txt
pandas
numpy
ccxt
scikit-learn
xgboost
statsmodels
arch
pmdarima
plotly
streamlit
cachetools
python-dotenv
